pyproject.toml
[project]
name = "deepseekdesktop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nuitka>=2.8.9",
    "openai>=2.14.0",
    "pyside6>=6.10.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "zstandard>=0.25.0",
]
